Completed checkpoint 3.
The main code file is 'checkpoint3.py'. I have added the additional file named 'has_image.py' which tells the co-ordinate of the star located in the 'checkpoint3.jpg' image and the second co-ordinate is detected in the image 'checkpoint3_copy.jpg'. The image of the star which has to be passed in the 'has_image.py' file has also been added.